test: add tests for content app

+++ content/views.py
@@ -1,15 +1,32 @@
 from datetime import timedelta
 
-from django.db.models import F, Case, When, Value, Model, Exists, OuterRef, Count, Q, Max
+from django.db.models import (
+    F,
+    Case,
+    When,
+    Value,
+    Model,
+    Exists,
+    OuterRef,
+    Count,
+    Q,
+    Max,
+)
 from django.db.models.fields import BooleanField
 from django.utils import timezone
 from rest_framework import viewsets, status
 from rest_framework.decorators import action
+from rest_framework.exceptions import ValidationError
 from rest_framework.permissions import IsAuthenticated
 from rest_framework.response import Response
 from content.models import Content, Follow, ContentEditHistory
 from notification.tasks import create_notification_for_followers_content
-from content.serializers import ContentSerializer, FollowSerializer, ContentAdminSerializer, FollowCreateSerializer
+from content.serializers import (
+    ContentSerializer,
+    FollowSerializer,
+    ContentAdminSerializer,
+    FollowCreateSerializer,
+)
 from django.core.cache import cache
 
 
@@ -19,25 +36,38 @@ class ContentViewSet(viewsets.ModelViewSet):
     def get_queryset(self):
         if not self.request.user.is_authenticated:
             return Content.objects.none()
-        queryset = Content.objects.select_related("author").prefetch_related("edit_history").annotate(
-            is_followed_by_me=Exists(
-                Follow.objects.filter(content=OuterRef("pk"), user=self.request.user)
-            ),
-            followers_count=Count(F("followers"), distinct=True),
-            is_hot=Case(
-                When(Q(followers_count__gt=5) | Q(edited_count__gt=10), then=Value(True)),
-                default=False,
-                output_field=BooleanField()
-            ),
-            recent_edits_count=Count("edit_history", filter=Q(edit_history__edited_at__gte=timezone.now() - timedelta(days=7))),
-            last_edit_date=Max("edit_history__edited_at")
+        queryset = (
+            Content.objects.select_related("author")
+            .prefetch_related("edit_history")
+            .annotate(
+                is_followed_by_me=Exists(
+                    Follow.objects.filter(
+                        content=OuterRef("pk"), user=self.request.user
+                    )
+                ),
+                followers_count=Count(F("followers"), distinct=True),
+                is_hot=Case(
+                    When(
+                        Q(followers_count__gt=5) | Q(edited_count__gt=10),
+                        then=Value(True),
+                    ),
+                    default=False,
+                    output_field=BooleanField(),
+                ),
+                recent_edits_count=Count(
+                    "edit_history",
+                    filter=Q(
+                        edit_history__edited_at__gte=timezone.now() - timedelta(days=7)
+                    ),
+                ),
+                last_edit_date=Max("edit_history__edited_at"),
+            )
         )
 
         if self.request.user.is_staff:
             queryset = queryset.prefetch_related("followers")
 
         return queryset
-
 
     def get_serializer_class(self):
 
@@ -47,10 +77,7 @@ class ContentViewSet(viewsets.ModelViewSet):
 
     def perform_create(self, serializer):
         new_content = serializer.save(author=self.request.user)
-        Follow.objects.create(
-            user=self.request.user,
-            content=new_content
-        )
+        Follow.objects.create(user=self.request.user, content=new_content)
 
     def retrieve(self, request, *args, **kwargs):
         content = self.get_object()
@@ -60,15 +87,20 @@ class ContentViewSet(viewsets.ModelViewSet):
             follow.last_viewed_at = timezone.now()
             follow.save(update_fields=["last_viewed_at"])
 
-        cache.delete_pattern(f"updated_contents_user_{request.user.pk}_page_*")
+        if hasattr(cache, "delete_pattern"):
+            cache.delete_pattern(f"updated_contents_user_{request.user.pk}_page_*")
+        else:
+            cache.clear()
         return super().retrieve(request, *args, **kwargs)
-
 
     def update(self, request, *args, **kwargs):
         content = self.get_object()
 
         if request.user != content.author:
-            return Response({"error": "You can only update your own posts."}, status=status.HTTP_403_FORBIDDEN)
+            return Response(
+                {"error": "You can only update your own posts."},
+                status=status.HTTP_403_FORBIDDEN,
+            )
 
         response = super().update(request, *args, **kwargs)
         create_notification_for_followers_content.delay(content.pk)
@@ -79,7 +111,10 @@ class ContentViewSet(viewsets.ModelViewSet):
         content = self.get_object()
 
         if request.user != content.author:
-            return Response({"error": "You can only delete your own posts."}, status=status.HTTP_403_FORBIDDEN)
+            return Response(
+                {"error": "You can only delete your own posts."},
+                status=status.HTTP_403_FORBIDDEN,
+            )
         return super().destroy(request, *args, **kwargs)
 
     def perform_update(self, serializer):
@@ -91,9 +126,11 @@ class ContentViewSet(viewsets.ModelViewSet):
 
         instance.edited_count = F("edited_count") + 1
         instance.save(update_fields=["edited_count"])
-        cache.delete_pattern(f"updated_contents_user_{self.request.user.pk}_page_*")
 
-
+        if hasattr(cache, "delete_pattern"):
+            cache.delete_pattern(f"updated_contents_user_{self.request.user.pk}_page_*")
+        else:
+            cache.clear()
 
     @action(methods=["GET"], url_path="updated-contents", detail=False)
     def update_contents(self, request):
@@ -106,21 +143,23 @@ class ContentViewSet(viewsets.ModelViewSet):
             return Response(cache_data, status=status.HTTP_200_OK)
 
         updated_follows = Follow.objects.filter(
-            user=request.user,
-            content__updated_at__gt=F("last_viewed_at")
+            user=request.user, content__updated_at__gt=F("last_viewed_at")
         )
 
         updated_contents = Content.objects.filter(
             followers__in=updated_follows
         ).order_by("-updated_at")
 
-
         page = self.paginate_queryset(updated_contents)
         if page is not None:
-            serializer = ContentSerializer(page, many=True, context=self.get_serializer_context())
+            serializer = ContentSerializer(
+                page, many=True, context=self.get_serializer_context()
+            )
             response_data = self.get_paginated_response(serializer.data).data
         else:
-            serializer = ContentSerializer(updated_contents, many=True, context=self.get_serializer_context())
+            serializer = ContentSerializer(
+                updated_contents, many=True, context=self.get_serializer_context()
+            )
             response_data = serializer.data
 
         cache.set(cache_key, response_data, timeout=300)
@@ -128,25 +167,37 @@ class ContentViewSet(viewsets.ModelViewSet):
         return Response(response_data, status=status.HTTP_200_OK)
 
 
-
 class FollowViewSet(viewsets.ModelViewSet):
+    http_method_names = ["get", "post", "delete", "head", "options"]
+
     def destroy(self, request, *args, **kwargs):
         follow = self.get_object()
         if follow.user != request.user and request.user.is_staff == False:
-            return Response({"error": "You can delete only your owns follows"}, status=status.HTTP_403_FORBIDDEN)
+            return Response(
+                {"error": "You can delete only your owns follows"},
+                status=status.HTTP_403_FORBIDDEN,
+            )
         return super().destroy(request, *args, **kwargs)
 
     def perform_create(self, serializer):
+        content = serializer.validated_data.get("content")
+
+        if Follow.objects.filter(user=self.request.user, content=content).exists():
+            raise ValidationError({"detail": "You are already following this content!"})
+
         serializer.save(user=self.request.user)
 
     def get_queryset(self):
         queryset = Follow.objects.select_related("user", "content").annotate(
-            has_new_changes = Case(
-                When(content__updated_at__gt=F("last_viewed_at"),
-                     then=Value(True)), default=Value(False),
-                     output_field=BooleanField()
+            has_new_changes=Case(
+                When(content__updated_at__gt=F("last_viewed_at"), then=Value(True)),
+                default=Value(False),
+                output_field=BooleanField(),
             ),
-            missed_edits_count=Count("content__edit_history", filter=Q(content__edit_history__edited_at__gt=F("last_viewed_at")))
+            missed_edits_count=Count(
+                "content__edit_history",
+                filter=Q(content__edit_history__edited_at__gt=F("last_viewed_at")),
+            ),
         )
 
         if self.request.user and self.request.user.is_staff:
